Add gameboard IIFE with method to place x on grid
I've created an IIFE for the gameboard object that
returns three methods creating an API that lets us
place an x within a 2d array, get the 2d array as
an object and print the 2d array as a table to the
console. This will give me visual feedback for how
the code is functioning. I've created the
gameboard as an IIFE as it only needs to be called
once and doesn't need to create instances of
itself like a factorial function or constructor.