Permit x or o to be entered only in empty grid pos
The gameboard IIFE now prompts for a marker 'x' or
'o', and only permits it to be added to an empty
position on the board. Additionally the board is
now contructed as a 2D array through a for loop
as opposed to being a fixed array. This will allow
me to later populate the board with 'cell' objects
that store the marker values.