Fill gameboard with cell objects redo printBoard
The gameboard is now built from cell objects that
come from the createCell factorial function. When
this was changed the printBoard method of our
gameboard now no longer displayed the cell values
in our console.table, instead of showed a table
full of objects - much less than ideal!
I had to rethink the printBoard method to just
return a table with the values of the cells
instead of entire cell objects. This fix is less
than perfect and might benefit from being mapped?