Fix placeMarker method of gameBoard
Our gameBoard was being populated with multiple
references to the same object. When I went to
update the values stored in the board it would
update every position of the board to the same
value. This was fixed by placing the cell variable
within the for loop's for loop - unique cells are
now being created in each position of the grid.